requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
pandas==2.1.4
APScheduler==3.10.4
requests==2.31.0
Werkzeug==3.0.1
pytest==8.0.0
pytest-mock==3.12.0
Flask-Limiter==3.5.0
celery==5.3.6
redis==5.0.1
python-json-logger==2.0.7
